Fixes extra creation of indirect sink edge for blocks with unresolved control flow
Cleans up compiler warnings that occur on Linux and not on Windows

Conflicts:
	dyninstAPI/src/Patching.C
	dyninstAPI/src/Relocation/CFG/RelocGraph.C
	dyninstAPI/src/Relocation/CFG/RelocGraph.h
	dyninstAPI/src/Relocation/Widgets/CFWidget-x86.C
	dyninstAPI/src/block.C

+++ dyninstAPI_RT/src/RTmemEmulator.c
@@ -214,7 +214,7 @@ int RTuntranslatedEntryCounter;
 extern void DYNINST_stopThread (void * pointAddr, void *callBackID, void *flags, void *calculation);
 
 void RThandleShadow(void *direction, void *pointAddr, void *callbackID, void *flags, void *calculation) {
-    if ((int)direction == 1) {
+    if ((long)direction == 1) {
         if (RTuntranslatedEntryCounter == 0) {
             // Entering a system call...
             DYNINST_stopThread(pointAddr, callbackID, flags, (void *)1);
